workspace restructure
